% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.to.tree.R
\name{convert.to.tree}
\alias{convert.to.tree}
\title{Create a phylogenetic tree from an ultrametric}
\usage{
convert.to.tree(n, L, u)
}
\arguments{
\item{n}{is the number of leaves}

\item{L}{is a vector of labels (strings) of leaves}

\item{u}{is an ultrametric}
}
\value{
A phylogenetic tree of class \code{phylo}
}
\description{
This function constructs a phylogenetic tree from an ultrametric.
}
\examples{
um<-Sim_Trees21[1,]
ll <- 10
L <- LETTERS[1:10]
tr<-convert.to.tree(ll, L, um)
}
\author{
Ruriko Yoshida \email{ryoshida@nps.edu}
}
