% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phylo.Gen.trees.T2.R
\docType{data}
\name{D2}
\alias{D2}
\title{Phylogenetic trees data simulated from the Coalescant model}
\format{
## `D2`
A 1000 x 45 matrix where each row is an ultrametric representing a phylogenetic tree.
}
\usage{
D2
}
\description{
1000 gene trees simulated from the Coalescant model based on a specified species
}
\keyword{datasets}
