% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.tpolytope.R
\name{draw.tpolytope}
\alias{draw.tpolytope}
\alias{draw.tpolytope.3d}
\title{Draw a 3-D tropical polytope}
\usage{
draw.tpolytope.3d(D, c)
}
\arguments{
\item{D}{Matrix of vertices of a tropical polytope.  Rows are the vertices.}

\item{c}{Color to render the polytope.}
}
\value{
3-D rendering of a tropical polytope.
}
\description{
This command draws a three dimensional tropical polytope.
}
\examples{
D <-matrix(c(0,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1),4,4,TRUE)
c<-'blue'

draw.tpolytope.3d(D,c)

}
\author{
Ruriko Yoshida \email{ryoshida@nps.edu}
}
