% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FWpoint.numerical.R
\name{FWpoint.numerical}
\alias{FWpoint.numerical}
\title{Fermat-Weber point numerical solver}
\usage{
FWpoint.numerical(datamatrix)
}
\arguments{
\item{datamatrix}{matrix of dimension N*e, where N is the number of
observations which lie in R^e.}
}
\value{
Fermat-Weber point approximation (vector in R^e)
}
\description{
Returns the Fermat-Weber point of N points using a gradient based numerical
method
}
\examples{
D = matrix(c(0,0,0,0,2,5,0,3,1),3,3,TRUE)
FWpoint.numerical(D)
}
\references{
Aliatimis, Georgios, Ruriko Yoshida, Burak Boyaci and James A.
Grant (2023). Tropical Logistic Regression on Space of Phylogenetic Trees
}
\author{
Georgios Aliatimis \email{g.aliatimis@lancaster.ac.uk}
}
