% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.ultrametric.R
\name{kernel.ultrametric}
\alias{kernel.ultrametric}
\title{Kernel Function for Ultrametrics.}
\usage{
kernel.ultrametric(x, mu, s)
}
\arguments{
\item{x}{Numeric vector representing a sampled ultrametric.}

\item{mu}{Numeric vector representing the centroid ultrametric.}

\item{s}{Bandwidth parameter based on tropical distance.}
}
\value{
Density value based on sampled point x.
}
\description{
This function calculates a non-parametric density for ultrametrics using a Gaussian kernel in conjunction with a tropical distance.
}
\author{
Ruriko Yoshida \email{ryoshida@nps.edu}
}
