% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lungfish_labels.R
\docType{data}
\name{lf_labels}
\alias{lf_labels}
\title{Leaf labels for lung fish data}
\format{
\subsection{\code{lf_labels}}{

A vector of species labels associated with the 1290 lung fish phylogenetic tree.
}
}
\source{
Liang D, Shen XX, Zhang P. One thousand two hundred ninety nuclear genes from a genome-wide survey support lungfishes as the sister group of tetrapods. Mol Biol Evol. 2013 Aug;30(8):1803-7. doi: 10.1093/molbev/mst072. Epub 2013 Apr 14. PMID: 23589454.
}
\usage{
lf_labels
}
\description{
Leaf labels associated with the 1290 (non-equidistant) gene trees originating from lung fish data in matrix form
}
\keyword{datasets}
