% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lung_fish_trees.R
\docType{data}
\name{Lung_Fish}
\alias{Lung_Fish}
\title{Phylogenetic trees based on lung fish data}
\format{
\subsection{\code{Lung_Fish}}{

A 1290 x 45 matrix where each row is an ultrametric representing a phylogenetic tree
}
}
\usage{
Lung_Fish
}
\description{
1290 (non-equidistant) gene trees originating from lung fish data in matrix form
}
\keyword{datasets}
