% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.ins.ball.R
\name{max_ins.ball}
\alias{max_ins.ball}
\title{Calculate the center point and radius of the maximum inscribed ball for a
tropical simplex}
\usage{
max_ins.ball(A, add = "max")
}
\arguments{
\item{A}{matrix of points defining a tropical polytope; rows are the points}

\item{add}{string; 'max' indicates max-plus addition, 'min' indicates
min-plus addition. Defaults to 'max'}
}
\value{
list containing the radius and center point of a maximum inscribed
ball
}
\description{
This function calculates the center point and radius of the maximum inscribed
ball for a max- or min-plus tropical simplex
}
\examples{
P<-matrix(c(0,0,0,0,2,5,0,3,1),3,3,TRUE)
max_ins.ball(P)
max_ins.ball(P,add='min')
}
\references{
Barnhill, David, Ruriko Yoshida and Keiji Miura (2023). Maximum
Inscribed and Minimum Enclosing Tropical Balls of Tropical Polytopes and
Applications to Volume Estimation and Uniform Sampling.
}
\author{
David Barnhill \email{david.barnhill@nps.edu}
}
