% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_enc_ball.R
\name{min_enc_ball}
\alias{min_enc_ball}
\title{Calculate a minimum enclosing ball for a tropical polytope.}
\usage{
min_enc_ball(A)
}
\arguments{
\item{A}{Matrix of points defining a tropical polytope. Rows are the points.}
}
\value{
List containing center point and radius of minimum enclosing ball of P.
}
\description{
This function constructs a minimum enclosing ball for a set of points defining a tropical polytope.
}
\examples{
P <-matrix(c(0,0,0,0,3,1,0,2,5),3,3,TRUE)
min_enc_ball(P)
}
\author{
David Barnhill \email{david.barnhill@nps.edu}
}
