% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliz.tree.R
\name{normaliz.tree}
\alias{normaliz.tree}
\title{Normalize a phylogenetic tree}
\usage{
normaliz.tree(D, h = 1)
}
\arguments{
\item{D}{numeric vector; ultrametric equidistant tree}

\item{h}{desired height; defaults to 1}
}
\value{
normalized equidistant tree
}
\description{
This function normalizes the height of a phylogenetic tree
}
\examples{
D <-c(4,4,2)
normaliz.tree(D, h=1)
}
\author{
Ruriko Yoshida \email{ryoshida@nps.edu}
}
