% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trop.clus.bet.KM.R
\name{over_bet_KM}
\alias{over_bet_KM}
\title{Tropical cluster betweeness measure for a each of a set of k-means derived set of clusters}
\usage{
over_bet_KM(A, C)
}
\arguments{
\item{A}{matrix of tropical points; rows are points with the last column representing a numbered cluster assignment}

\item{C}{number of clusters}
}
\value{
betweenness cluster measure
}
\description{
This function calculates an overall betweenness measure between a set of clusters derived from tropical k-means clustering
}
\examples{
P<-matrix(c(0,-100,-100,0,100,0,0,0,100),3,3,TRUE)
x0<-c(0,0,0)
y0<-c(0,0,0)
z0<-c(0,0,0)
mu1<-c(0,-10,-20)
sig1<-5
mu2<-c(0,20,30)
sig2<-8
mu3<-c(0,30,10)
sig3<-3
N<-50
har_norms1<-matrix(0,N,3)
har_norms2<-matrix(0,N,3)
har_norms3<-matrix(0,N,3)

for(i in 1:N){
  print(i)
  x<-TropicalPolytope.extrapolation.HAR_NORM(P, x0, I = 50,mu1,sig1)
  y<-TropicalPolytope.extrapolation.HAR_NORM(P, y0, I = 50,mu2,sig2)
  z<-TropicalPolytope.extrapolation.HAR_NORM(P, z0, I = 50,mu3,sig3)
  x0<-x
  y0<-y
  z0<-z
  har_norms1[i,]<-x
  har_norms2[i,]<-y
  har_norms3[i,]<-z
}

hars<-rbind(har_norms1,har_norms2,har_norms3)
cls<-c(rep(1,50),rep(2,50),rep(3,50))
cl_pt<-cbind(hars,cls)

C<-3
over_bet_KM(cl_pt,C)
}
\references{
David Barnhill, Ruriko Yoshida (2023). Clustering Methods Over the Tropically Convex Sets.
}
\author{
David Barnhill \email{david.barnhill@nps.edu}
}
