% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Points.TLineSeg.R
\name{Points.TLineSeg}
\alias{Points.TLineSeg}
\title{Sample k equally spaced points on a max-plus tropical line segment}
\usage{
Points.TLineSeg(D1, D2, k = 20)
}
\arguments{
\item{D1}{Point in the tropical projective torus}

\item{D2}{Point in the tropical projective torus}

\item{k}{Number of points}
}
\value{
A matrix of k equally spaced points on a tropical line segment.
}
\description{
This function calculates k equally spaced points on a tropical line segment.
}
\examples{
D1 <-c(0,4,2)
D2 <- c(0,7,-1)
p<-Points.TLineSeg(D1, D2, k = 5)
}
\author{
Ruriko Yoshida \email{ryoshida@nps.edu}
}
