% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{real number}
}
\value{
sigmoid function value at x
}
\description{
Returns the sigmoid function valuation
}
\examples{
sigmoid(0) # 0.5
}
\references{
Aliatimis, Georgios, Ruriko Yoshida, Burak Boyaci and James A.
Grant (2023). Tropical Logistic Regression on Space of Phylogenetic Trees
}
\author{
Georgios Aliatimis \email{g.aliatimis@lancaster.ac.uk}
}
