% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLineSeg_min.R
\name{TLineSeg_min}
\alias{TLineSeg_min}
\title{Construct a min-plus tropical line segment accor}
\usage{
TLineSeg_min(D1, D2)
}
\arguments{
\item{D1}{point in the tropical projective torus}

\item{D2}{point in the tropical projective torus}
}
\value{
list of points defining the tropical line segment
}
\description{
This function constructs a min-plus tropical line segment between two points
}
\examples{
D1 <-c(0,4,2)
D2 <- c(0,7,-1)
TLineSeg_min(D1, D2)
}
\author{
David Barnhill \email{david.barnhill@nps.edu}
}
