% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TML-package.R
\docType{package}
\name{TML-package}
\alias{TML-package}
\title{TML: Tropical Geometry Tools for Machine Learning}
\description{
Suite of tropical geometric tools for use in machine learning applications. These methods may be summarized in the following references: Yoshida, et al. (2022) \doi{10.2140/astat.2023.14.37}, Barnhill et al. (2023) \doi{10.48550/arXiv.2303.02539}, Barnhill and Yoshida (2023) \doi{10.3390/math11153433}, Aliatimis et al. (2023) \doi{10.1007/s11538-024-01327-8}, Yoshida et al. (2022) \doi{10.1109/TCBB.2024.3420815}, and Yoshida et al. (2019) \doi{10.1007/s11538-018-0493-4}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/barnhilldave/TML}
  \item Report bugs at \url{https://github.com/barnhilldave/TML/issues}
}

}
\author{
\strong{Maintainer}: David Barnhill \email{david.barnhill@nps.edu} (\href{https://orcid.org/0009-0006-2781-7434}{ORCID})

Authors:
\itemize{
  \item Ruriko Yoshida
  \item Georgios Aliatimis
  \item Keiji Miura
}

}
\keyword{internal}
