% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trop.min.hyper.dist.R
\name{trop.dist.hyp_min}
\alias{trop.dist.hyp_min}
\title{Calculate the tropical distance to a min-tropical hyperplane}
\usage{
trop.dist.hyp_min(O, x0)
}
\arguments{
\item{O}{normal vector of a tropical hyperplane; numeric vector}

\item{x0}{point of interest; numeric vector}
}
\value{
tropical distance to min-plus tropical hyperplane
}
\description{
Calculate the tropical distance to a min-tropical hyperplane
}
\examples{
O <-c(0,-1,-1)
x0 <- c(0,-2,-8)
trop.dist.hyp_min(O,x0)
}
\author{
David Barnhill \email{david.barnhill@nps.edu}
}
