% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trop_dist.R
\name{trop.dist}
\alias{trop.dist}
\title{Compute the tropical distance}
\usage{
trop.dist(D1, D2)
}
\arguments{
\item{D1}{Point in the tropical projective torus}

\item{D2}{Point in the tropical projective torus}
}
\value{
Tropical distance between D1 and D2
}
\description{
This function computes the tropical distance between two points in the tropical projective torus
}
\examples{
D1 <-c(0,4,2)
D2 <- c(0,7,-1)
trop.dist(D1, D2)

}
\author{
Ruriko Yoshida \email{ryoshida@nps.edu}
}
