% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trop.SVM.R
\name{trop.SVM}
\alias{trop.SVM}
\title{Tropical support vector machine}
\usage{
trop.SVM(V, clses, L = 1000, tr = 0.8, st = 100, trop = 1)
}
\arguments{
\item{V}{matrix of data points where each row is an observation; does not include dependent variables}

\item{clses}{vector of classifications of each observation (response variable)}

\item{L}{maximum number of iterations to use}

\item{tr}{proportion of data used for training set}

\item{st}{number of iterations to terminate MCMC if current state does not change}

\item{trop}{logical; 1 if data possesses a leading column of zeros; 0 otherwise}
}
\value{
list with confusion matrix and correct classification rate
}
\description{
Supervised learning using tropical support vector machines in conjunction with simulated annealing
}
\examples{
\dontrun{
library(HDclassif)
wine1<-as.matrix(wine)
wines<-wine1[,2:ncol(wine1)]
w_type<-wine1[,1]
trop.SVM(wines,w_type,L=1000,tr=.8,st=100,trop=1)
}
}
\references{
Gärtner, Bernd & Jaggi, Martin (2023). Tropical Support Vector Machines.

Ruriko Yoshida, Misaki Takamori, Hideyuki Matsumoto, Keiji Miura (2023).Tropical support vector machines: Evaluations and extension to function spaces,
Neural Networks, Volume 157, Pages 77-89, ISSN 0893-6080.
}
\author{
David Barnhill \email{david.barnhill@nps.edu}
}
