% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropical.HC.AGNES.R
\name{Tropical.HC.AGNES}
\alias{Tropical.HC.AGNES}
\title{Agglomerative (AGNES) tropical hierarchical clustering}
\usage{
Tropical.HC.AGNES(D, method = "average")
}
\arguments{
\item{D}{Matrix of points defining a tropical polytope.  Rows are the tropical points.}

\item{method}{Linkage method: "average", "min", or "max".}
}
\value{
List of distances in when merges occur; list of indices of points in each cluster.
}
\description{
This function performs agglomerative (AGNES) hierarchical clustering over the space of ultrametrics defining the space of equidistant trees.
}
\examples{
P <-Sim_points
Tropical.HC.AGNES(P, method="average")
}
\author{
David Barnhill \email{david.barnhill@nps.edu}
}
