% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropical.HC.DIANA.R
\name{Tropical.HC.DIANA}
\alias{Tropical.HC.DIANA}
\title{Divisive (DIANA) tropical hierarchical clustering}
\usage{
Tropical.HC.DIANA(D, D1, D2, n, k = 2)
}
\arguments{
\item{D}{matrix of ultrametrics; rows are the tropical points}

\item{D1}{subset of D}

\item{D2}{subset of D}

\item{n}{number of leaves}

\item{k}{number of clusters to partition data}
}
\value{
value
}
\description{
This function performs divisive (DIANA) hierarchical clustering over the space of ultrametrics defining the space of equidistant trees.
}
\examples{
set.seed(123)
index.1 <- sample(nrow(Sim_Trees1),30)
index.2 <- sample(nrow(Sim_Trees2),30)
T1<-Sim_Trees1[index.1,]
T2<-Sim_Trees2[index.2,]
T<-rbind(T1,T2)
n=10
Tropical.HC.DIANA(T,T1,T2, n,k=2)
}
\references{
David Barnhill, Ruriko Yoshida (2023). Clustering Methods Over the Tropically Convex Sets.
}
\author{
David Barnhill \email{david.barnhill@nps.edu}
}
