% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lung.fish.trees.R
\docType{data}
\name{lung_fish}
\alias{lung_fish}
\title{Phylogenetic trees based on lung fish data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1290 rows and 45 columns.
}
\source{
Liang D, Shen XX, Zhang P. One thousand two hundred ninety nuclear
genes from a genome-wide survey support lungfishes as the sister group of
tetrapods. Mol Biol Evol. 2013 Aug;30(8):1803-7. doi:
10.1093/molbev/mst072. Epub 2013 Apr 14. PMID: 23589454.
}
\usage{
lung_fish
}
\description{
1290 (non-equidistant) gene trees with 45 leaves originating from lung fish
data in matrix form.  Also we provide a vector of strings consisting of leaf
labels for each species associated with the data set.
}
\keyword{dataset}
\keyword{lungfish}
