% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.points.NORM.R
\docType{data}
\name{Sim_points}
\alias{Sim_points}
\title{Simulated points over the tropical projective torus}
\format{
\subsection{\code{Sim_points}}{

A 150 x 3 matrix where each row is a point in the tropical projective torus
}
}
\usage{
Sim_points
}
\description{
150 points generated using Gaussian-like Hit-and-Run sampling with three separate pairs of location and scale parameters
}
\keyword{datasets}
