% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLineSeg.R
\name{TLineSeg}
\alias{TLineSeg}
\title{Construct a max-plus tropical line segment accor}
\usage{
TLineSeg(D1, D2)
}
\arguments{
\item{D1}{Point in the tropical projective torus}

\item{D2}{Point in the tropical projective torus}
}
\value{
A list of points defining the tropical line segment.
}
\description{
This function constructs a max-plus tropical line segment between two points.
}
\examples{
D1 <-c(0,4,2)
D2 <- c(0,7,-1)
TLineSeg(D1, D2)
}
\author{
Ruriko Yoshida \email{ryoshida@nps.edu}
}
