% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TML-package.R
\docType{package}
\name{TML}
\alias{TML}
\alias{package-TML}
\title{Tropical Machine Learning in R}
\description{
TML provides a suite of tools for machine learning application on data over the tropical semiring
}
