% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropical.FW.R
\name{Trop_FW}
\alias{Trop_FW}
\title{Calculate the tropical Fermat-Weber point}
\usage{
Trop_FW(A)
}
\arguments{
\item{A}{Matrix of points defining a tropical polytope.  Rows are the tropical points.}
}
\value{
Numeric vector providing the tropical Fermat-Weber point for the tropical polytope.
}
\description{
This function calculates the Fermat-Weber point for a tropical polytope.
}
\examples{
P <-matrix(c(0,0,0,0,2,5,0,3,1),3,3,TRUE)

Trop_FW(P)

}
\author{
David Barnhill \email{david.barnhill@nps.edu}
}
