% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trop_Volume.R
\name{Trop_Volume}
\alias{Trop_Volume}
\title{Estimate the volume of a tropical polytope.}
\usage{
Trop_Volume(B, P, x0, S, I, R)
}
\arguments{
\item{B}{Matrix of points defining a minimum enclosing ball for a polytope P. Rows are the points.}

\item{P}{Matrix of points defining a tropical polytope. Rows are the points.}

\item{x0}{Initial point used for the HAR sampler.}

\item{S}{Number of points to sample from the minimum enclosing ball.}

\item{I}{Number of iterations for the HAR sampler.}

\item{R}{Radius of the minimum enclosing tropical ball.}
}
\value{
List containing ratio of points falling in P; volume of the tropical ball; volume estimate of P.
}
\description{
This function uses tropical HAR with a uniform target distribution to estimate the volume of a tropical polytope.
}
\examples{
P <-matrix(c(0,0,0,0,3,1,0,2,5),3,3,TRUE)
BR<-min_enc_ball(P)
B<-trop_bal.vert(BR[[1]],BR[[2]])
x0<-c(0,1.5,.4)
S<-500
I<-50
R<-BR[[2]]
Trop_Volume(B,P,x0,S,I,R)
}
\author{
David Barnhill \email{david.barnhill@nps.edu}
}
