% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ultrametrics.HAR.R
\name{Ultrametrics.HAR}
\alias{Ultrametrics.HAR}
\title{Hit-and-Run Sampler for the space of ultrametrics}
\usage{
Ultrametrics.HAR(x0, n, I = 1, h = 1)
}
\arguments{
\item{x0}{A phylogenetic tree}

\item{n}{Number of leaves for the equidistant tree.}

\item{I}{Number of states in the Markov chain.}

\item{h}{Height of phylogenetic tree.}
}
\value{
Point in the space of ultrametrics over n leaves.
}
\description{
This sampler samples a point in the space of ultrametrics where each point represents an equidistant tree on n leaves.
}
\examples{
x0 <-Sim_Trees1[1,]
n<-10

Ultrametrics.HAR(x0, n, I = 50, h = 1)

}
\author{
Ruriko Yoshida \email{ryoshida@nps.edu}
}
