% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_to_equidistant_tree.R
\name{vector.to.equidistant.tree}
\alias{vector.to.equidistant.tree}
\title{Vector to equidistant tree}
\usage{
vector.to.equidistant.tree(omega)
}
\arguments{
\item{omega}{vector of pairwise distances in R^(m choose 2), where m is the number of leaves}
}
\value{
equidistant phylogenetic tree
}
\description{
A vector of pairwise distances is used to reconstruct the corresponding equidistant tree
}
\examples{
omega = c(1/3,1,1,1,1,1/3)
tree = vector.to.equidistant.tree(omega)
plot(tree)
}
\references{
Aliatimis, Georgios, Ruriko Yoshida, Burak Boyaci and James A. Grant (2023). Tropical Logistic Regression on Space of Phylogenetic Trees
}
\author{
Georgios Aliatimis \email{g.aliatimis@lancaster.ac.uk}
}
